+++ src/api/Models/Borehole.cs
@@ -455,6 +455,16 @@ public class Borehole : IChangeTracking, IIdentifyable
     public ICollection<Completion> Completions { get; }
 
     /// <summary>
+    /// Gets the <see cref="Borehole"/>'s sections.
+    /// </summary>
+    public ICollection<Section> Sections { get; }
+
+    /// <summary>
+    /// Gets the <see cref="Borehole"/>'s observations.
+    /// </summary>
+    public ICollection<Observation> Observations { get; }
+
+    /// <summary>
     /// Gets the <see cref="Borehole"/>'s workflows.
     /// </summary>
     public ICollection<Workflow> Workflows { get; } = new List<Workflow>();