+++ tests/Controllers/BoreholeControllerTest.cs
@@ -1,6 +1,7 @@
 ﻿using BDMS.Authentication;
 using BDMS.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
@@ -70,11 +71,11 @@ public class BoreholeControllerTest
 
         Assert.AreNotSame(originalStratigraphy.ChronostratigraphyLayers, copiedstratigraphy.ChronostratigraphyLayers);
         Assert.AreNotEqual(originalStratigraphy.ChronostratigraphyLayers.First().Id, copiedstratigraphy.ChronostratigraphyLayers.First().Id);
-        Assert.AreEqual(originalStratigraphy.ChronostratigraphyLayers.First().ChronostratigraphyId, copiedstratigraphy.ChronostratigraphyLayers.First().ChronostratigraphyId);
+        Assert.AreEqual(originalStratigraphy.ChronostratigraphyLayers.OrderBy(c => c.Id).First().ChronostratigraphyId, copiedstratigraphy.ChronostratigraphyLayers.OrderBy(c => c.Id).First().ChronostratigraphyId);
 
         Assert.AreNotSame(originalStratigraphy.LithostratigraphyLayers, copiedstratigraphy.LithostratigraphyLayers);
         Assert.AreNotEqual(originalStratigraphy.LithostratigraphyLayers.First().Id, copiedstratigraphy.LithostratigraphyLayers.First().Id);
-        Assert.AreEqual(originalStratigraphy.LithostratigraphyLayers.First().LithostratigraphyId, copiedstratigraphy.LithostratigraphyLayers.First().LithostratigraphyId);
+        Assert.AreEqual(originalStratigraphy.LithostratigraphyLayers.OrderBy(l => l.Id).First().LithostratigraphyId, copiedstratigraphy.LithostratigraphyLayers.OrderBy(l => l.Id).First().LithostratigraphyId);
 
         Assert.AreNotSame(originalBorehole.BoreholeFiles, copiedBorehole.BoreholeFiles);
         Assert.AreNotEqual(originalBorehole.BoreholeFiles.First().BoreholeId, copiedBorehole.BoreholeFiles.First().BoreholeId);
@@ -89,6 +90,47 @@ public class BoreholeControllerTest
 
         Assert.AreNotSame(originalStratigraphy.Layers.First().LayerUscs3Codes, copiedstratigraphy.Layers.First().LayerUscs3Codes);
         Assert.AreEqual(originalStratigraphy.Layers.First().LayerUscs3Codes.Count, copiedstratigraphy.Layers.First().LayerUscs3Codes.Count);
+
+        var originalCompletion = originalBorehole.Completions.First();
+        var copiedCompletion = copiedBorehole.Completions.First();
+
+        Assert.AreNotEqual(originalCompletion.Id, copiedCompletion.Id);
+        Assert.AreNotSame(originalCompletion.Casings, copiedCompletion.Casings);
+        Assert.AreNotEqual(originalCompletion.Casings.First().Id, copiedCompletion.Casings.First().Id);
+        Assert.AreEqual(originalCompletion.Casings.First().Notes, copiedCompletion.Casings.First().Notes);
+
+        Assert.AreNotSame(originalCompletion.Backfills, copiedCompletion.Backfills);
+        Assert.AreNotEqual(originalCompletion.Backfills.First().Id, copiedCompletion.Backfills.First().Id);
+        Assert.AreEqual(originalCompletion.Backfills.First().Created, copiedCompletion.Backfills.First().Created);
+
+        Assert.AreNotSame(originalCompletion.Instrumentations, copiedCompletion.Instrumentations);
+        Assert.AreNotEqual(originalCompletion.Instrumentations.First().Id, copiedCompletion.Instrumentations.First().Id);
+        Assert.AreEqual(originalCompletion.Instrumentations.First().Status, copiedCompletion.Instrumentations.First().Status);
+
+        Assert.AreNotEqual(originalCompletion.Casings.First().CasingElements.First().Id, copiedCompletion.Casings.First().CasingElements.First().Id);
+        Assert.AreEqual(originalCompletion.Casings.First().CasingElements.OrderBy(c => c.Id).First().OuterDiameter, copiedCompletion.Casings.First().CasingElements.OrderBy(c => c.Id).First().OuterDiameter);
+
+        var originalWaterIngress = originalBorehole.Observations.OfType<WaterIngress>().First();
+        var copiedWaterIngress = copiedBorehole.Observations.OfType<WaterIngress>().First();
+
+        Assert.AreNotEqual(originalWaterIngress.Id, copiedWaterIngress.Id);
+        Assert.AreEqual(originalWaterIngress.IsOpenBorehole, copiedWaterIngress.IsOpenBorehole);
+
+        var originalFieldMeasurement = originalBorehole.Observations.OfType<FieldMeasurement>().First();
+        var copiedFieldMeasurement = copiedBorehole.Observations.OfType<FieldMeasurement>().First();
+
+        Assert.AreNotEqual(originalFieldMeasurement.Id, copiedFieldMeasurement.Id);
+        Assert.AreNotSame(originalFieldMeasurement.FieldMeasurementResults, copiedFieldMeasurement.FieldMeasurementResults);
+        Assert.AreNotEqual(originalFieldMeasurement.FieldMeasurementResults.First().Id, copiedFieldMeasurement.FieldMeasurementResults.First().Id);
+        Assert.AreEqual(originalFieldMeasurement.FieldMeasurementResults.First().Value, copiedFieldMeasurement.FieldMeasurementResults.First().Value);
+
+        var originalSection = originalBorehole.Sections.First();
+        var copiedSection = copiedBorehole.Sections.First();
+
+        Assert.AreNotEqual(originalSection.Id, copiedSection.Id);
+        Assert.AreNotSame(originalSection.SectionElements, copiedSection.SectionElements);
+        Assert.AreNotEqual(originalSection.SectionElements.First().Id, copiedSection.SectionElements.First().Id);
+        Assert.AreEqual(originalSection.SectionElements.First().FromDepth, copiedSection.SectionElements.First().FromDepth);
     }
 
     private Borehole GetBorehole(int id)
@@ -99,8 +141,22 @@ public class BoreholeControllerTest
     // Get the id of a borehole with certain conditions.
     private int GetBoreholeIdToCopy()
     {
-        var borehole = GetBoreholesWithIncludes(context.Boreholes)
+        List<Borehole> boreholes = GetBoreholesWithIncludes(context.Boreholes).ToList();
+
+        foreach (var bh in boreholes)
+        {
+           context.Entry(bh)
+                .Collection(b => b.Observations)
+                .Query()
+                .OfType<FieldMeasurement>()
+                .Include(f => f.FieldMeasurementResults)
+                .Load();
+        }
+
+        var borehole = boreholes
             .Where(b =>
+                b.Stratigraphies != null &&
+                b.Stratigraphies.Any() &&
                 b.Stratigraphies.First().Layers != null &&
                 b.Stratigraphies.First().Layers.Any(x => x.LayerColorCodes != null && x.LayerColorCodes.Any()) &&
                 b.Stratigraphies.First().Layers.Any(x => x.LayerGrainShapeCodes != null && x.LayerGrainShapeCodes.Any()) &&
@@ -109,6 +165,14 @@ public class BoreholeControllerTest
                 b.Stratigraphies.First().FaciesDescriptions != null &&
                 b.Stratigraphies.First().ChronostratigraphyLayers != null &&
                 b.Stratigraphies.First().LithostratigraphyLayers != null &&
+                b.Completions.First() != null &&
+                b.Completions.First().Casings.First() != null &&
+                b.Completions.First().Casings.First().CasingElements.First() != null &&
+                b.Observations.First() != null &&
+                b.Observations.OfType<WaterIngress>().Any() &&
+                b.Observations.OfType<FieldMeasurement>().Any(fm => fm.FieldMeasurementResults.Count != 0) &&
+                b.Sections.First() != null &&
+                b.Sections.First().SectionElements.First() != null &&
                 b.BoreholeFiles.First().File != null &&
                 b.Canton != null &&
                 b.Stratigraphies.First().ChronostratigraphyLayers.First().ChronostratigraphyId != null)